+++ packages/terra-dropdown-button/src/terra-dev-site/test/dropdown-button/DefaultSplitButton.test.jsx
@@ -7,10 +7,10 @@ const cx = classnames.bind(styles);
 
 export default () => (
   <div className={cx('container-spacing-wrapper')}>
-    <SplitButton primaryOptionLabel="Split" onSelect={() => {}} id="split">
-      <Item label="1st" onSelect={() => {}} />
-      <Item label="2nd" onSelect={() => {}} />
-      <Item label="3rd" onSelect={() => {}} />
+    <SplitButton primaryOptionLabel="Split" metaData={{ key: 'primary-button' }} onSelect={(event, metaData) => { console.log('Metdata of :', metaData.key); }} id="split">
+      <Item label="1st" metaData={{ key: '1st-option' }} onSelect={(event, metaData) => { console.log('Metdata of :', metaData.key); }} />
+      <Item label="2nd" metaData={{ key: '2nd-option' }} onSelect={(event, metaData) => { console.log('Metdata of :', metaData.key); }} />
+      <Item label="3rd" metaData={{ key: '3rd-option' }} onSelect={(event, metaData) => { console.log('Metdata of :', metaData.key); }} />
     </SplitButton>
   </div>
 );